# Conflicts:
#	.github/workflows/pull_request_ubuntu_build.yml
#	.github/workflows/pull_request_windows_build.yml
#	compiler/ballerina-lang/src/main/java/io/ballerina/projects/BuildOptions.java
#	compiler/ballerina-lang/src/main/java/io/ballerina/projects/CompilationOptions.java
#	compiler/ballerina-lang/src/main/java/org/ballerinalang/compiler/CompilerOptionName.java
#	compiler/ballerina-lang/src/main/java/org/ballerinalang/util/diagnostic/DiagnosticErrorCode.java
#	compiler/ballerina-lang/src/main/java/org/wso2/ballerinalang/compiler/semantics/analyzer/ConstantValueResolver.java
#	compiler/ballerina-lang/src/main/java/org/wso2/ballerinalang/compiler/semantics/analyzer/SymbolEnter.java

+++ compiler/ballerina-lang/src/main/java/org/wso2/ballerinalang/compiler/semantics/model/symbols/BStructureTypeSymbol.java
@@ -40,7 +40,7 @@ public abstract class BStructureTypeSymbol extends BTypeSymbol {
     //todo remove once BTypeReferenceType is introduced to runtime
     public BTypeDefinitionSymbol typeDefinitionSymbol;
 
-    BStructureTypeSymbol(SymbolKind kind, int symTag, long flags, Name name, PackageID pkgID, BType type,
+    BStructureTypeSymbol(SymbolKind kind, long symTag, long flags, Name name, PackageID pkgID, BType type,
                          BSymbol owner, Location pos, SymbolOrigin origin) {
         super(symTag, flags, name, pkgID, type, owner, pos, origin);
         this.attachedFuncs = new ArrayList<>(0);
